Bug 1073003 - Fix -Wnarrowing warnings. r=ehsan
UltraBlame original commit: 6550cdd34f6fce661f2fb72da4974013449ae701

+++ dom/media/omx/OMXCodecDescriptorUtil.cpp
@@ -84,8 +84,8 @@ struct AVCParamSet {
   {
     
     uint8_t size[] = {
-      (mSize & 0xFF00) >> 8, 
-      mSize & 0x00FF,        
+      uint8_t((mSize & 0xFF00) >> 8), 
+      uint8_t(mSize & 0x00FF),        
     };
     aOutputBuf->AppendElements(size, sizeof(size));
